+++ src/assets/images/index.ts
@@ -1,4 +1,6 @@
 import HotelImage from "./Rectangle 19.png";
+import Logo from "./logo.png";
 export const imageList = {
   HotelImage,
+  Logo,
 };