+++ src/components/atoms/Breadcrumb/index.tsx
@@ -1,50 +1,28 @@
-import {
-  BreadcrumbItem,
-  Breadcrumb,
-  BreadcrumbLink,
-  Text,
-} from "@chakra-ui/react";
-
-import { useNavigate } from "react-router-dom";
 import { RxHamburgerMenu } from "react-icons/rx";
 import styled from "styled-components";
 import { colors } from "@soaltee-loyalty/theme/colors";
 import { getSidebarState } from "@soaltee-loyalty/components/organisms/layout";
-const Divider = styled.div`
-  color: ${colors.light_gray};
-  margin-left: 10px;
-`;
+
 const Wrapper = styled.div`
   display: flex;
   align-items: center;
   gap: 15px;
+  span {
+    line-height: 17px;
+    color: ${colors.text_black};
+  }
 `;
-const BreadCrumbContent = styled(Breadcrumb)``;
 interface IBreadCrumb {
-  items: { name: string; route: string }[];
-  goBack?: string;
+  name: string;
 }
 
-export const BreadCrumb = ({ items }: IBreadCrumb) => {
-  // const { t } = useTranslation();
-  const navigate = useNavigate();
+export const BreadCrumb = ({ name }: IBreadCrumb) => {
   const { showSidebar, setShowSidebar } = getSidebarState();
 
   return (
     <Wrapper>
       <RxHamburgerMenu onClick={() => setShowSidebar(!showSidebar)} />
-      <BreadCrumbContent spacing={1} separator={""}>
-        {items.map((item, i) => (
-          <BreadcrumbItem key={i}>
-            <BreadcrumbLink onClick={() => navigate(item.route)}>
-              <Text fontWeight={700} color={colors.primary}>
-                {item.name}
-              </Text>
-            </BreadcrumbLink>
-            {items.length - 1 !== i && <Divider>/</Divider>}
-          </BreadcrumbItem>
-        ))}
-      </BreadCrumbContent>
+      <span>{name}</span>
     </Wrapper>
   );
 };