+++ src/components/templates/master-data/property-list/index.tsx
@@ -1,5 +1,4 @@
 import { Stack, useDisclosure } from "@chakra-ui/react";
-import { BreadCrumb } from "@soaltee-loyalty/components/atoms/Breadcrumb";
 import { ProductForm } from "@soaltee-loyalty/components/templates/form";
 import ModalForm from "@soaltee-loyalty/components/organisms/modal";
 import DataTable, {
@@ -11,7 +10,7 @@ import { useGetProducts } from "@soaltee-loyalty/service/service-list";
 import { useMemo, useState } from "react";
 import { CellProps } from "react-table";
 
-const ListingPage = () => {
+const PropertyList = () => {
   const { data: tableData, isFetching: tableDataFetching } = useGetProducts();
   const [, setUpdateId] = useState("");
 
@@ -24,7 +23,6 @@ const ListingPage = () => {
   } = useDisclosure();
   const { isOpen, onOpen, onClose } = useDisclosure();
 
-  // Pagination
   const [pageParams, setPageParams] = useState({
     page: 1,
     limit: 10,
@@ -38,8 +36,6 @@ const ListingPage = () => {
   };
   const _pageSizeChange = (limit: number) =>
     setPageParams({ ...pageParams, limit, page: 1 });
-
-  // Pagination ends
 
   const columns = useMemo(
     () => [
@@ -59,27 +55,18 @@ const ListingPage = () => {
       },
       {
         Header: "Action",
+        width: "100px",
         Cell: ({ row }: CellProps<{ id: string; name: string }>) => {
-          // const onEdit = () => {
-          //   setUpdateId(row.original?.id);
-          //   // setIsUpdate(true);
-          //   onProductModalOpen();
-          // };
           const onView = () => {
             setUpdateId(row.original?.id);
             onViewProductModalOpen();
           };
           const onDelete = () => {
             onOpen();
-            // setbankID(row?.original?.id);
           };
           return (
             <Stack alignItems={"flex-start"}>
-              <TableActions
-                // onEdit={onEdit}
-                onView={onView}
-                onDelete={onDelete}
-              />
+              <TableActions onView={onView} onDelete={onDelete} />
             </Stack>
           );
         },
@@ -90,7 +77,6 @@ const ListingPage = () => {
   console.log(paginatedData);
   return (
     <>
-      <BreadCrumb name="List" />
       <DataTable
         data={paginatedData || []}
         loading={tableDataFetching}
@@ -110,20 +96,14 @@ const ListingPage = () => {
       />
       <ModalForm
         isModalOpen={isProductOpen}
-        // isLoading={isLoading || isUpdating || bankInfoFetching}
-        // title={isUpdate ? "Update Bank" : "Add Bank"}
         onCloseModal={onProductModalClose}
         resetButtonText={"Cancel"}
-        // submitButtonText={isUpdate ? "Update Bank" : "Add bank"}
-        // submitHandler={handleSubmit(onSubmitHandler)}
       >
         <p>Add Product</p>
       </ModalForm>
 
-      {/* view modal */}
       <ModalForm
         isModalOpen={isViewProductOpen}
-        // isLoading={ProductInfoFetching}
         title="Product Details"
         onCloseModal={onViewProductModalClose}
         submitButtonText={"Okay"}
@@ -132,20 +112,16 @@ const ListingPage = () => {
         <p>View Modal</p>
       </ModalForm>
 
-      {/* view modal ends */}
-
       <ModalForm
         title={"Delete"}
-        // isLoading={isDeleting}
         isModalOpen={isOpen}
         onCloseModal={onClose}
         resetButtonText={"No"}
         submitButtonText={"Yes"}
-        // handleSubmit={() => onDelete(bankID ?? "")}
       >
         Are you sure you want to delete the Product detail?
       </ModalForm>
     </>
   );
 };
-export default ListingPage;
+export default PropertyList;