+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## 0.0.14 (2023-08-28)
+## 0.0.14 (2023-08-29)
 
 
 ### Bug Fixes
@@ -10,7 +10,6 @@
 ### Features
 
 * add additional folder ([6a3e80e](https://gitlab.com/paras205/boilerplate-react/commit/6a3e80e0283bb533435461c064b30935cd953ddf))
-* added signup component ([10f4fb5](https://gitlab.com/paras205/boilerplate-react/commit/10f4fb5a3680ca0affd8164c8c1493ce816a75f3))
 * custom input ([d323c8f](https://gitlab.com/paras205/boilerplate-react/commit/d323c8f84f00f51287e6c564fa41478402ce5ae9))
 * login component ([f753932](https://gitlab.com/paras205/boilerplate-react/commit/f753932002845eaab1196b776d992ee5ecac0547))
 * login component ([967e159](https://gitlab.com/paras205/boilerplate-react/commit/967e159a4fe973d5e6302f108f2bd08525fcc615))